+++ scripts/External Integration Clustering/case2_clustering.py
@@ -5,7 +5,7 @@ This script demonstrates a class based approach for clustering with KMeans Using
 import argparse
 
 from relevanceai import Client
-from relevanceai.clusterer import Clusterer
+from relevanceai.clusterer import ClusterOps
 from relevanceai.clusterer.kmeans_clusterer import KMeansModel
 
 
@@ -18,7 +18,7 @@ def main(args):
 
     model = KMeansModel(k=3)
 
-    clusterer = Clusterer(model=model, alias=f"kmeans-{n_clusters}")
+    clusterer = ClusterOps(model=model, alias=f"kmeans-{n_clusters}")
 
     clusterer.fit(dataset=df, vector_fields=[vector_field])